+++ source/kmarius_incremental_scan/changelog.md
@@ -1,3 +1,7 @@
+**<span style="color:#56adda">0.4.0</span>**
+- add misc. settings for the panel
+- panel: improve loading times for large libraries
+
 **<span style="color:#56adda">0.3.0</span>**
 - send status messages to the ui during scan